feat(projects): интеграция бека со списком проекта (#32)
Jira issue: VEGA-483
