+++ codegen.yml
@@ -1,0 +1,15 @@
+overwrite: true
+schema:
+  - "schema.graphql":
+      headers:
+        Authorization: Bearer ${AUTH_TOKEN}
+documents: "src/**/*.ts"
+generates:
+  src/generated/graphql.tsx:
+    plugins:
+      - "typescript"
+      - "typescript-operations"
+      - "typescript-react-apollo"
+  ./graphql.schema.json:
+    plugins:
+      - "introspection"
