+++ src/data/projects/index.ts
@@ -1,0 +1,1 @@
+export { GET_PROJECTS_QUERY } from './query';
