+++ src/layouts/DataLayout/index.ts
@@ -1,0 +1,1 @@
+export { DataLayout } from './DataLayout';
